Day-18/pyproject.toml
[project]
name = "day-18"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "colorgram-py==1.2.0",
]
